update pipeline for optimum intel
